Rename MessageIDAdapter to MessageIdAdapter
